Refactoring: Created a module (security) for shared Spring Security JWT components
Initializing a parent Maven project for shared components
