+++ pom.xml
@@ -37,5 +37,6 @@
 	<modules>
 		<module>infrastructure</module>
 		<module>services</module>
+		<module>shared</module>
 	</modules>
 </project>
