+++ services/notification-service/src/main/java/ru/dreadblade/stockmarket/notificationservice/config/SecurityConfiguration.java
@@ -6,8 +6,8 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.security.web.SecurityFilterChain;
-import ru.dreadblade.stockmarket.notificationservice.security.StockMarketJwtAuthenticationConverter;
-import ru.dreadblade.stockmarket.notificationservice.security.StockMarketJwtDecoder;
+import ru.dreadblade.stockmarket.shared.security.StockMarketJwtAuthenticationConverter;
+import ru.dreadblade.stockmarket.shared.security.StockMarketJwtDecoder;
 
 @Configuration
 @EnableWebSecurity
