+++ services/order-service/pom.xml
@@ -11,7 +11,6 @@
 	<artifactId>order-service</artifactId>
 	<name>order-service</name>
 	<description>order-service</description>
-	<version>0.0.1-SNAPSHOT</version>
 
 	<dependencies>
 		<dependency>
@@ -33,16 +32,9 @@
 		</dependency>
 
 		<dependency>
-			<groupId>org.springframework.boot</groupId>
-			<artifactId>spring-boot-starter-security</artifactId>
-		</dependency>
-		<dependency>
-			<groupId>org.springframework.security</groupId>
-			<artifactId>spring-security-oauth2-resource-server</artifactId>
-		</dependency>
-		<dependency>
-			<groupId>org.springframework.security</groupId>
-			<artifactId>spring-security-oauth2-jose</artifactId>
+			<groupId>ru.dreadblade.stockmarket.shared</groupId>
+			<artifactId>security</artifactId>
+			<version>${project.version}</version>
 		</dependency>
 
 		<dependency>
