+++ services/pom.xml
@@ -12,7 +12,6 @@
     <artifactId>services</artifactId>
     <name>services</name>
     <description>services</description>
-    <version>0.0.1-SNAPSHOT</version>
 
     <packaging>pom</packaging>
