+++ services/stock-price-history-service/pom.xml
@@ -12,10 +12,6 @@
 	<name>stock-price-history-service</name>
 	<description>stock-price-history-service</description>
 
-	<properties>
-		<java.version>17</java.version>
-	</properties>
-
 	<dependencies>
 		<dependency>
 			<groupId>org.springframework.boot</groupId>
