+++ services/stock-price-history-service/src/main/java/ru/dreadblade/stockmarket/stockpricehistoryservice/api/mapper/StockPriceHistoryMapper.java
@@ -2,7 +2,7 @@ package ru.dreadblade.stockmarket.stockpricehistoryservice.api.mapper;
 
 import org.springframework.stereotype.Component;
 import ru.dreadblade.stockmarket.stockpricehistoryservice.api.model.StockPriceHistoryResponseDTO;
-import ru.dreadblade.stockmarket.stockpricehistoryservice.domain.StockPriceHistory;
+import ru.dreadblade.stockmarket.stockpricehistoryservice.model.StockPriceHistory;
 
 @Component
 public class StockPriceHistoryMapper {
