+++ services/stock-price-history-service/src/main/java/ru/dreadblade/stockmarket/stockpricehistoryservice/repository/StockRepository.java
@@ -1,7 +1,7 @@
 package ru.dreadblade.stockmarket.stockpricehistoryservice.repository;
 
 import org.springframework.data.mongodb.repository.MongoRepository;
-import ru.dreadblade.stockmarket.stockpricehistoryservice.domain.Stock;
+import ru.dreadblade.stockmarket.stockpricehistoryservice.model.Stock;
 
 public interface StockRepository extends MongoRepository<Stock, Long> {
 }
