+++ services/stock-price-service/pom.xml
@@ -11,11 +11,6 @@
 	<artifactId>stock-price-service</artifactId>
 	<name>stock-price-service</name>
 	<description>stock-price-service</description>
-	<version>0.0.1-SNAPSHOT</version>
-
-	<properties>
-		<java.version>17</java.version>
-	</properties>
 
 	<dependencies>
 		<dependency>
