+++ services/stock-service/pom.xml
@@ -11,11 +11,6 @@
 	<artifactId>stock-service</artifactId>
 	<name>stock-service</name>
 	<description>stock-service</description>
-	<version>0.0.1-SNAPSHOT</version>
-
-	<properties>
-		<java.version>17</java.version>
-	</properties>
 
 	<dependencies>
 		<dependency>
@@ -45,16 +40,9 @@
 		</dependency>
 
 		<dependency>
-			<groupId>org.springframework.boot</groupId>
-			<artifactId>spring-boot-starter-security</artifactId>
-		</dependency>
-		<dependency>
-			<groupId>org.springframework.security</groupId>
-			<artifactId>spring-security-oauth2-resource-server</artifactId>
-		</dependency>
-		<dependency>
-			<groupId>org.springframework.security</groupId>
-			<artifactId>spring-security-oauth2-jose</artifactId>
+			<groupId>ru.dreadblade.stockmarket.shared</groupId>
+			<artifactId>security</artifactId>
+			<version>${project.version}</version>
 		</dependency>
 
 		<dependency>
