+++ shared/security/pom.xml
@@ -4,41 +4,28 @@
 	<modelVersion>4.0.0</modelVersion>
 	<parent>
 		<groupId>ru.dreadblade.stockmarket</groupId>
-		<artifactId>services</artifactId>
+		<artifactId>shared</artifactId>
 		<version>0.0.1-SNAPSHOT</version>
 	</parent>
 
-	<artifactId>user-service</artifactId>
-	<name>user-service</name>
-	<description>user-service</description>
-
-	<properties>
-		<keycloak.version>19.0.0</keycloak.version>
-	</properties>
+	<groupId>ru.dreadblade.stockmarket.shared</groupId>
+	<name>security</name>
+	<description>security</description>
+	<artifactId>security</artifactId>
+	<version>0.0.1-SNAPSHOT</version>
 
 	<dependencies>
 		<dependency>
 			<groupId>org.springframework.boot</groupId>
-			<artifactId>spring-boot-starter-web</artifactId>
+			<artifactId>spring-boot-starter-security</artifactId>
 		</dependency>
 		<dependency>
-			<groupId>org.springframework.boot</groupId>
-			<artifactId>spring-boot-starter-actuator</artifactId>
-		</dependency>
-
-		<dependency>
-			<groupId>org.keycloak</groupId>
-			<artifactId>keycloak-admin-client</artifactId>
-			<version>${keycloak.version}</version>
-		</dependency>
-
-		<dependency>
-			<groupId>org.springframework.cloud</groupId>
-			<artifactId>spring-cloud-starter-netflix-eureka-client</artifactId>
+			<groupId>org.springframework.security</groupId>
+			<artifactId>spring-security-oauth2-resource-server</artifactId>
 		</dependency>
 		<dependency>
-			<groupId>org.springframework.cloud</groupId>
-			<artifactId>spring-cloud-starter-config</artifactId>
+			<groupId>org.springframework.security</groupId>
+			<artifactId>spring-security-oauth2-jose</artifactId>
 		</dependency>
 
 		<dependency>
@@ -55,5 +42,4 @@
 			</plugin>
 		</plugins>
 	</build>
-
 </project>
