+++ shared/security/src/main/java/ru/dreadblade/stockmarket/shared/security/StockMarketSecurityAutoconfiguration.java
@@ -1,0 +1,10 @@
+package ru.dreadblade.stockmarket.shared.security;
+
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+@ComponentScan("ru.dreadblade.stockmarket.shared.security")
+public class StockMarketSecurityAutoconfiguration {
+
+}
